feat: Update crypto balance display
This commit updates the display of the crypto balance in the `Balance` component and the `MainBalance` component. The `Balance` component now receives the `values` prop from the `MainBalance` component, which contains the current and last amounts of the crypto balance. The default values for `current_amount` and `last_amount` in the `Balance` component have been changed to 0.0. Additionally, the `Home` component now retrieves the crypto balance values from the global store and uses them to update the displayed balance when the "Show" button is clicked.

Note: This commit message follows the established conventions in the repository.